Adds some code to read out an object from the Store and return it (B)
M	svc/storage/file_backend.py

This adds a get_object method which extracts the content from an
object in the store and converts it into an appropriate
Object. However, the on disk file doesn't know it's own type so we
need to make changes to our infrastrcture for far. This is a bad
commit and we need to fix this after we build the rest of the format
changes.